feat add getPlayer

+++ script/HelperConfig.s.sol
@@ -22,7 +22,7 @@ contract HelperConfig is Script, CodeConstants {
         uint256 interval;
         address vrfCoordinator;
         bytes32 keyHash;
-        uint64 subscriptionId;
+        uint256 subscriptionId;
         uint32 callbackGasLimit;
     }
